python/requirements.txt
jsonschema<3.0.0
flasgger==0.9.3
Flask==1.1.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
pyyaml==5.4.1
python-dateutil
